Packages installed. Webpack config is now working as expected
