+++ client/src/index.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+// import $ from 'jquery';
+// import axios from 'axios';
+
+class App extends React.Component {
+  constructor() {
+    super();
+    this.state = {
+
+    }
+  }
+
+
+  render() {
+    return (
+      <div>donna</div>
+    );
+  }
+
+}
+
+ReactDOM.render(<App />, document.getElementById('app'));
