--- client/src/js/components/container/App.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-// import PropTypes from "prop-types";
-
-class App extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-    };
-  }
-
-  render() {
-    return (
-      <div>donna</div>
-    );
-  }
-
-}
-
-ReactDOM.render(<App />, document.getElementById('app'));
