+++ package.json
@@ -8,11 +8,12 @@
     "node": ">=6.13.0"
   },
   "scripts": {
-    "start": "webpack-dev-server --open --mode development",
+    "start": "nodemon server/index.js",
     "build": "webpack --mode production"
   },
   "dependencies": {
-    "mongoose": "^5.4.10"
+    "mongoose": "^5.4.10",
+    "mongoose-auto-increment": "^5.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
