+++ server/index.js
@@ -3,7 +3,7 @@ const db = require('../database-mongodb/index.js');
 const Product = require('../database-mongodb/Product.js');
 
 const app = express();
-const PORT = 3000;
+const PORT = 3002;
 
 
 app.use(express.static(__dirname + '/../client/dist'));
@@ -15,4 +15,5 @@ app.get('/api/product:id/sizechart', function(req, res) {
 
 app.listen(PORT, () => {
   console.log(`listening on port ${PORT}`);
+  console.log(__dirname + '/../client/dist');
 });
